+++ src/test-common/route-check-parameters.js
@@ -1,34 +1,52 @@
-function defineSingleNodeObject(inpNode, graphNodes)
+function defineSingleNodeObject(inpNode, inpGraphNodes)
 {
 	var defineRes = {};
-	defineRes["targetNode"] = graphNodes.indexOf(inpNode);
+	defineRes["targetNode"] = inpGraphNodes.indexOf(inpNode);
 	return defineRes;
 }
 
 
-function defineMultipleNodesObject(inpNodeList, graphNodes)
+function defineMultipleNodesObject(inpNodeList, inpGraphNodes)
+{
+	var defineRes = {};
+	defineRes["targetList"] = mapNodes(inpNodeList, inpGraphNodes);
+	return defineRes;
+}
+
+function defineRouteTypeObject(inpStartNodes, inpEndNodes, inpGraphNodes)
+{
+	var defineRes = {};
+	
+	defineRes["start"] = mapNodes(inpStartNodes, inpGraphNodes);
+	defineRes["end"] = mapNodes(inpEndNodes, inpGraphNodes);
+	
+	return defineRes;
+}
+
+
+
+function mapNodes(inputNodes, graphNodes)
 {
 	var itemIndex = 0;
 	var currentItem = null;
 	var currentNodeIndex = -1;
 	var currentUsed = false;
 	
-	var defineRes = {};
-	defineRes["targetList"] = [];
+	var mapRes = [];
 	
-	for (itemIndex = 0; itemIndex < inpNodeList.length; itemIndex = itemIndex + 1)
+	for (itemIndex = 0; itemIndex < inputNodes.length; itemIndex = itemIndex + 1)
 	{
-		currentItem = inpNodeList[itemIndex];
+		currentItem = inputNodes[itemIndex];
 		currentNodeIndex = graphNodes.indexOf(currentItem);
-		currentUsed = defineRes.targetList.includes(currentNodeIndex);
+		currentUsed = mapRes.includes(currentNodeIndex);
 		
 		if (currentNodeIndex >= 0 && currentNodeIndex < graphNodes.length && currentUsed !== true)
 		{
-			defineRes.targetList.push(currentNodeIndex);
+			mapRes.push(currentNodeIndex);
 		}
 	}
 	
-	return defineRes;
+	return mapRes;
 }
 
 
@@ -36,5 +54,6 @@ function defineMultipleNodesObject(inpNodeList, graphNodes)
 module.exports =
 {
 	defineSingleNode: defineSingleNodeObject,
-	defineMultipleNodes: defineMultipleNodesObject
+	defineMultipleNodes: defineMultipleNodesObject,
+	defineRouteType: defineRouteTypeObject
 };