+++ test-parts/test-possible-routes-valid.js
@@ -18,6 +18,7 @@ function runTests()
 		handleOneWay();
 		handleStartNode();
 		handleEndNode();
+		//handleRouteTypes();
 		//handleArguments();
 		//handleRouteCount();
 	});
@@ -121,6 +122,106 @@ function handleEndNode()
 			var resultParas = routeCheckParameters.defineMultipleNodes("ABCDEFGH", exampleGraphObject.nodes);
 			
 			routeResults.checkObject(resultValue, exampleGraphObject, testScenarios.END_MULT, resultParas);
+		});
+		
+	});
+}
+
+
+function handleRouteTypes()
+{
+	describe("Route Types", function()
+	{
+		it("One-To-One", function()
+		{
+			var routeStart = routeCriteria.defineStartNode("A");
+			var routeEnd = routeCriteria.defineEndNode("C");
+			var routeOneWay = routeCriteria.defineOneWay();
+			
+			var searchCriteria = [routeStart, routeEnd, routeOneWay];
+			var resultValue = possibleRoutes.findRoutes(exampleGraphObject, searchCriteria);
+			var resultParas = routeCheckParameters.defineRouteType("A", "C", exampleGraphObject.nodes);
+			
+			routeResults.checkObject(resultValue, exampleGraphObject, testScenarios.ROUTE_TYPE, resultParas);
+		});
+		
+		it("One-To-Multiple", function()
+		{
+			var routeStart = routeCriteria.defineStartNode("A");
+			var routeEndD = routeCriteria.defineEndNode("D");
+			var routeEndE = routeCriteria.defineEndNode("E");
+			var routeOneWay = routeCriteria.defineOneWay();
+			
+			var searchCriteria = [routeStart, routeEndD, routeEndE, routeOneWay];
+			var resultValue = possibleRoutes.findRoutes(exampleGraphObject, searchCriteria);
+			var resultParas = routeCheckParameters.defineRouteType("A", "DE", exampleGraphObject.nodes);
+			
+			routeResults.checkObject(resultValue, exampleGraphObject, testScenarios.ROUTE_TYPE, resultParas);
+		});
+		
+		it("One-To-All", function()
+		{
+			var routeStart = routeCriteria.defineStartNode("C");
+			var routeOneWay = routeCriteria.defineOneWay();
+			
+			var searchCriteria = [routeStart, routeOneWay];
+			var resultValue = possibleRoutes.findRoutes(exampleGraphObject, searchCriteria);
+			var resultParas = routeCheckParameters.defineRouteType("C", "ABCDEFGH", exampleGraphObject.nodes);
+			
+			routeResults.checkObject(resultValue, exampleGraphObject, testScenarios.ROUTE_TYPE, resultParas);
+		});
+		
+		it("Multiple-To-One", function()
+		{
+			var routeStartA = routeCriteria.defineStartNode("A");
+			var routeStartB = routeCriteria.defineStartNode("B");
+			var routeStartC = routeCriteria.defineStartNode("C");
+			var routeEnd = routeCriteria.defineEndNode("F");
+			var routeOneWay = routeCriteria.defineOneWay();
+			
+			var searchCriteria = [routeStartA, routeStartB, routeStartC, routeEnd, routeOneWay];
+			var resultValue = possibleRoutes.findRoutes(exampleGraphObject, searchCriteria);
+			var resultParas = routeCheckParameters.defineRouteType("ABC", "F", exampleGraphObject.nodes);
+			
+			routeResults.checkObject(resultValue, exampleGraphObject, testScenarios.ROUTE_TYPE, resultParas);
+		});
+		
+		it("Multiple-To-Multiple", function()
+		{
+			var routeStartA = routeCriteria.defineStartNode("A");
+			var routeStartB = routeCriteria.defineStartNode("B");
+			var routeEndC = routeCriteria.defineEndNode("C");
+			var routeEndD = routeCriteria.defineEndNode("D");
+			var routeOneWay = routeCriteria.defineOneWay();
+			
+			var searchCriteria = [routeStartA, routeStartB, routeEndC, routeEndD, routeOneWay];
+			var resultValue = possibleRoutes.findRoutes(exampleGraphObject, searchCriteria);
+			var resultParas = routeCheckParameters.defineRouteType("AB", "CD", exampleGraphObject.nodes);
+			
+			routeResults.checkObject(resultValue, exampleGraphObject, testScenarios.ROUTE_TYPE, resultParas);
+		});
+		
+		it("Multiple-To-All", function()
+		{
+			var routeStartA = routeCriteria.defineStartNode("A");
+			var routeStartB = routeCriteria.defineStartNode("B");
+			var routeOneWay = routeCriteria.defineOneWay();
+			
+			var searchCriteria = [routeStartA, routeStartB, routeOneWay];
+			var resultValue = possibleRoutes.findRoutes(exampleGraphObject, searchCriteria);
+			var resultParas = routeCheckParameters.defineRouteType("AB", "ABCDEFGH", exampleGraphObject.nodes);
+		});
+		
+		it("All-To-One", function()
+		{
+			var routeEnd = routeCriteria.defineEndNode("B");
+			var routeOneWay = routeCriteria.defineOneWay();
+			
+			var searchCriteria = [routeEnd, routeOneWay];
+			
+			var searchCriteria = [routeStartA, routeStartB, routeOneWay];
+			var resultValue = possibleRoutes.findRoutes(exampleGraphObject, searchCriteria);
+			var resultParas = routeCheckParameters.defineRouteType("ABCDEFGH", "B", exampleGraphObject.nodes);
 		});
 		
 	});